feat: add decoding multiple fields at once
When using `field()` and `field_optional()`, if one field throws an
error, none of the fields after that one are decoded. This is good
when consuming values from a source with a defined structure like an
API or database, but is less helpful when getting values from users
like form submissions. When a user submits a form, you want to let
them know of all problems at once instead of one at a time.

The new `field_group()` function checks all fields (Field instances)
passed to it and either returns the decoded values or an error for
each instead of stopping on the first problem.